Moved helper functions and precomputed values into utils.py
