+++ recognition/46429515_OASIS_brain_SD/module.py
@@ -1,26 +1,11 @@
+import utils
 import torch
 import torch.nn.functional as F
 from torch import nn
 import math
 
-device = torch.device('cuda' if torch.cuda.is_available() else 'cpu')
-
-
 # Noise Scheduler (Forward Process)
-
-def quadratic_beta_schedule(timesteps, start=0.0001, end=0.02):
-    return torch.linspace(start**0.5, end**0.5, timesteps)**2
-
-def get_index_from_list(values, t, x_shape):
-    """
-    Helper function that returns specific index of t of a list of values
-    vals while considering batch dimension
-    """
-    batch_size = t.shape[0]
-    out = values.gather(-1, t.cpu())
-    return out.reshape(batch_size, *((1,) * (len(x_shape) - 1))).to(t.device)
-
-def forward_diffusion_sample(x_0, t, device=device):
+def forward_diffusion_sample(x_0, t, device=utils.device):
     """
     Takes image and timestep as input to return the image w/ noise
     q(x_t|x_0) = N(x_t;sqrt(alpha_t)*x_0, (1-alpha_t)I)
@@ -28,31 +13,17 @@ def forward_diffusion_sample(x_0, t, device=device):
     x_t = sqrt(alpha_cumprod_t) * x_0 + sqrt(1 - alpha_cumprod_t) * noise
     """
     noise = torch.randn_like(x_0)
-    sqrt_alphas_cumprod_t = get_index_from_list(sqrt_alphas_cumprod, t, x_0.shape)
-    sqrt_one_minus_alphas_cumprod_t = get_index_from_list(sqrt_one_minus_alphas_cumprod, t, x_0.shape)
+    sqrt_alphas_cumprod_t = utils.get_index_from_list(utils.sqrt_alphas_cumprod, t, x_0.shape)
+    sqrt_one_minus_alphas_cumprod_t = utils.get_index_from_list(utils.sqrt_one_minus_alphas_cumprod, t, x_0.shape)
     
     # mean + variance
     return sqrt_alphas_cumprod_t.to(device) * x_0.to(device) \
         + sqrt_one_minus_alphas_cumprod_t.to(device) * noise.to(device), noise.to(device)
 
-# Define beta schedule
-T = 500
-betas = quadratic_beta_schedule(timesteps=T)
-
-# Pre-calculate different terms for closed form
-alphas = 1. - betas
-alphas_cumprod = torch.cumprod(alphas, axis=0)
-alphas_cumprod_prev = F.pad(alphas_cumprod[:-1], (1, 0), value=1.0)
-sqrt_recip_alphas = torch.sqrt(1.0 / alphas)
-sqrt_alphas_cumprod = torch.sqrt(alphas_cumprod)
-sqrt_one_minus_alphas_cumprod = torch.sqrt(1. - alphas_cumprod)
-posterior_variance = betas * (1. - alphas_cumprod_prev) / (1. - alphas_cumprod)
-
 
 # U-Net (Backwards Process)
 
 # Aim is to obtain x_0, from x_t
-
 class Block(nn.Module):
     def __init__(self, in_channel, out_channel, time_emb_dim, up=False):
         super().__init__()
