+++ recognition/46429515_OASIS_brain_SD/utils.py
@@ -20,6 +20,36 @@ IMAGE_SIZE = 128
 # Define batch size of data
 BATCH_SIZE = 32
 
+
+## module.py
+
+# Define beta schedule
+
+def quadratic_beta_schedule(timesteps, start=0.0001, end=0.02):
+    return torch.linspace(start**0.5, end**0.5, timesteps)**2
+
+def get_index_from_list(values, t, x_shape):
+    """
+    Helper function that returns specific index of t of a list of values
+    vals while considering batch dimension
+    """
+    batch_size = t.shape[0]
+    out = values.gather(-1, t.cpu())
+    return out.reshape(batch_size, *((1,) * (len(x_shape) - 1))).to(t.device)
+
+T = 500
+betas = quadratic_beta_schedule(timesteps=T)
+
+# Pre-calculate different terms for closed form
+alphas = 1. - betas
+alphas_cumprod = torch.cumprod(alphas, axis=0)
+alphas_cumprod_prev = F.pad(alphas_cumprod[:-1], (1, 0), value=1.0)
+sqrt_recip_alphas = torch.sqrt(1.0 / alphas)
+sqrt_alphas_cumprod = torch.sqrt(alphas_cumprod)
+sqrt_one_minus_alphas_cumprod = torch.sqrt(1. - alphas_cumprod)
+posterior_variance = betas * (1. - alphas_cumprod_prev) / (1. - alphas_cumprod)
+
+
 ## train.py
 
 # Device setup
@@ -51,17 +81,17 @@ def sample_timestep(model, x, t):
     """
     Noise in the image x is predicted and returns denoised image
     """
-    betas_t = module.get_index_from_list(module.betas, t, x.shape)
+    betas_t = get_index_from_list(betas, t, x.shape)
     sqrt_one_minus_alphas_cumprod_t = module.get_index_from_list(
-        module.sqrt_one_minus_alphas_cumprod, t, x.shape
+        sqrt_one_minus_alphas_cumprod, t, x.shape
     )
-    sqrt_recip_alphas_t = module.get_index_from_list(module.sqrt_recip_alphas, t, x.shape)
+    sqrt_recip_alphas_t = get_index_from_list(sqrt_recip_alphas, t, x.shape)
     
     # call model (current image - noise prediction)
     model_mean = sqrt_recip_alphas_t * (
         x - betas_t * model(x, t) / sqrt_one_minus_alphas_cumprod_t
     )
-    posterior_variance_t = module.get_index_from_list(module.posterior_variance, t, x.shape)
+    posterior_variance_t = get_index_from_list(posterior_variance, t, x.shape)
     
     if t == 0:
         return model_mean
@@ -93,7 +123,7 @@ def save_tensor_image(image, epoch, step, output_dir):
 @torch.no_grad()
 def sample_save_image(model, epoch, output_dir):
     # Sample noise
-    img_size = dataset.IMAGE_SIZE
+    img_size = IMAGE_SIZE
     img = torch.randn((1, 1, img_size, img_size), device=device)
     num_images = 10
     stepsize = int(module.T/num_images)
